syscallcompat: add asUser for darwin and make OpenatUser and friends shared
